Updated multiple notify messages

+++ PokeStar/PokeStar/ConnectionInterface/Connections.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 using System.Collections.Generic;
 using Discord;
@@ -28,8 +29,6 @@ namespace PokeStar.ConnectionInterface
       private Dictionary<int, List<string>> Eggs;
       private Dictionary<string, Rocket> Rockets;
 
-      private const int NumSuggestions = 10;
-
       /// <summary>
       /// Creates a new Connections object.
       /// Private to implement the singleton design patturn.
@@ -49,11 +48,6 @@ namespace PokeStar.ConnectionInterface
          if (connections == null)
          {
             connections = new Connections();
-            connections.UpdatePokemonNameList();
-            connections.UpdateMoveNameList();
-            connections.UpdateRaidBossList();
-            connections.UpdateEggList();
-            connections.UpdateRocketList();
          }
          return connections;
       }
@@ -196,9 +190,9 @@ namespace PokeStar.ConnectionInterface
 
          Dictionary<int, List<string>> newBosses = SilphData.GetRaidBosses();
 
-         bool bossesChanged = RaidBosses.Keys.Count != newBosses.Keys.Count || !RaidBosses.Keys.All(newBosses.Keys.Contains);
+         bool bossesChanged = RaidBosses == null || RaidBosses.Keys.Count != newBosses.Keys.Count || !RaidBosses.Keys.All(newBosses.Keys.Contains);
 
-         foreach (int tier in RaidBosses.Keys)
+         foreach (int tier in newBosses.Keys)
          {
             bossesChanged = bossesChanged || RaidBosses[tier].Count != newBosses[tier].Count || !RaidBosses[tier].All(newBosses[tier].Contains);
          }
@@ -210,33 +204,55 @@ namespace PokeStar.ConnectionInterface
                UpdateRaidBossList();
                SocketGuild emoteServer = guilds.FirstOrDefault(x => x.Name.Equals(Global.EMOTE_SERVER, StringComparison.OrdinalIgnoreCase));
                GuildEmote[] previousEmotes = emoteServer.Emotes.ToArray();
-
-               foreach (GuildEmote emote in emoteServer.Emotes)
-               {
-                  await emoteServer.DeleteEmoteAsync(emote);
-               }
+               bool msgChange = false;
 
                foreach (KeyValuePair<int, List<string>> tier in RaidBosses)
                {
                   foreach (string boss in tier.Value)
                   {
                      string fileName = GetPokemonPicture(boss);
-                     CopyFile(fileName);
-                     Image img = new Image(fileName);
-                     await emoteServer.CreateEmoteAsync(fileName.Remove(fileName.Length - 4), img);
-                     img.Dispose();
-                     DeleteFile(fileName);
+                     string emoteName = fileName.Remove(fileName.Length - 4);
+                     if (previousEmotes.FirstOrDefault(x => x.Name.Equals(emoteName)) == null)
+                     {
+                        CopyFile(fileName);
+                        Image img = new Image(fileName);
+                        await emoteServer.CreateEmoteAsync(emoteName, img);
+                        img.Dispose();
+                        DeleteFile(fileName);
+                        msgChange = true;
+                     }
                   }
                }
 
-               Dictionary<ulong, ulong> channels = Instance().GetNotificationChannels();
-
-               foreach (KeyValuePair<ulong, ulong> chan in channels)
+               foreach (GuildEmote emote in previousEmotes)
                {
-                  SocketGuild guild = guilds.FirstOrDefault(x => x.Id == chan.Key);
-                  ISocketMessageChannel channel = (ISocketMessageChannel)guild.Channels.FirstOrDefault(x => x.Id == chan.Value);
-                  await ClearNotifyMessage(guild, channel.Id, previousEmotes.ToArray());
-                  await SetNotifyMessage(guild, channel, emoteServer.Emotes.ToArray());
+                  bool inCurrentList = false;
+                  foreach (KeyValuePair<int, List<string>> tier in RaidBosses)
+                  {
+                     foreach (string boss in tier.Value)
+                     {
+                        string fileName = GetPokemonPicture(boss);
+                        string emoteName = fileName.Remove(fileName.Length - 4);
+                        inCurrentList = inCurrentList || emote.Name.Equals(emoteName);
+                     }
+                  }
+                  if (!inCurrentList)
+                  {
+                     await emoteServer.DeleteEmoteAsync(emote);
+                     msgChange = true;
+                  }
+               }
+               if (msgChange)
+               {
+                  Dictionary<ulong, ulong> channels = Instance().GetNotificationChannels();
+
+                  foreach (KeyValuePair<ulong, ulong> chan in channels)
+                  {
+                     SocketGuild guild = guilds.FirstOrDefault(x => x.Id == chan.Key);
+                     ISocketMessageChannel channel = (ISocketMessageChannel)guild.Channels.FirstOrDefault(x => x.Id == chan.Value);
+                     await ClearNotifyMessage(guild, channel, Instance().GetNotificationMessages(guild.Id, channel.Id));
+                     await SetNotifyMessage(guild, channel, emoteServer.Emotes.ToArray());
+                  }
                }
             }
          }
@@ -249,7 +265,7 @@ namespace PokeStar.ConnectionInterface
       /// <returns>True if the message is a raid notification message, otherwise false.</returns>
       public static bool IsNotifyMessage(ulong id)
       {
-         return Connections.Instance().CheckNotificationMessage(id);
+         return Instance().CheckNotificationMessage(id);
       }
 
       /// <summary>
@@ -300,13 +316,34 @@ namespace PokeStar.ConnectionInterface
             await guild.CreateRoleAsync(emote.Name, null, Global.ROLE_COLOR_RAID, false, true, null);
          }
 
-         IUserMessage message = await ResponseMessage.SendInfoMessage(channel,
-            "React to this message to be notified when a specific raid is called.\n" +
-            "When the raid bosses change your role will be removed and you will have to re-select desired bosses.\n" +
-            "If you no longer wish to be notified for a boss, re-react for the desired boss.");
-         message.AddReactionsAsync(emotes);
+         StringBuilder messages = new StringBuilder();
 
-         Instance().UpdateNotificationMessage(guild.Id, channel.Id, message.Id);
+         var initialMessage = await ResponseMessage.SendInfoMessage(channel,
+            "React to the following messages to be notified when a specific raid is called.\n" +
+            "When the raid bosses change, your role will be removed and you will have to re-select desired bosses.\n" +
+            "If you no longer wish to be notified for a boss, re-react for the desired boss.");
+         messages.AppendLine($"{initialMessage.Id},");
+
+         StringBuilder emoteMsg = new StringBuilder();
+
+         int count;
+         for (count = 0; count < +emotes.Length; count++)
+         {
+            emoteMsg.AppendLine($"{emotes[count]} {emotes[count].Name}");
+
+            if (count % Global.LIMIT_RAID_BOSS_NOTIFY == Global.LIMIT_RAID_BOSS_NOTIFY - 1)
+            {
+               IUserMessage emoteMessage = await ResponseMessage.SendInfoMessage(channel, emoteMsg.ToString());
+               await emoteMessage.AddReactionsAsync(emotes.Skip(count - (Global.LIMIT_RAID_BOSS_NOTIFY - 1)).Take(Global.LIMIT_RAID_BOSS_NOTIFY).ToArray());
+               messages.AppendLine($"{emoteMessage.Id},");
+               emoteMsg.Clear();
+            }
+         }
+         IUserMessage finalEmoteMessage = await ResponseMessage.SendInfoMessage(channel, emoteMsg.ToString());
+         await finalEmoteMessage.AddReactionsAsync(emotes.Skip(count - (count % Global.LIMIT_RAID_BOSS_NOTIFY)).Take(Global.LIMIT_RAID_BOSS_NOTIFY).ToArray());
+         messages.AppendLine($"{finalEmoteMessage.Id},");
+
+         Instance().UpdateNotificationMessage(guild.Id, channel.Id, messages.ToString().Remove(messages.Length - 1, 1));
       }
 
       /// <summary>
@@ -316,17 +353,20 @@ namespace PokeStar.ConnectionInterface
       /// <param name="channel">Id of registered channel.</param>
       /// <param name="emotes">Emotes that represent roles.</param>
       /// <returns></returns>
-      public static async Task ClearNotifyMessage(SocketGuild guild, ulong channel, Emote[] emotes)
+      public static async Task ClearNotifyMessage(SocketGuild guild, ISocketMessageChannel channel, List<ulong> messages)
       {
-         foreach (GuildEmote emote in emotes)
+         List<SocketRole> roles = guild.Roles.Where(x => x.Color.Equals(Global.ROLE_COLOR_RAID)).ToList();
+         foreach (SocketRole role in roles)
          {
-            SocketRole role = guild.Roles.FirstOrDefault(x => x.Color.Equals(Global.ROLE_COLOR_RAID));
-            if (role != null)
-            {
-               await role.DeleteAsync();
-            }
+            await role.DeleteAsync();
          }
-         Instance().ClearNotificationMessage(guild.Id, channel);
+
+         foreach (ulong message in messages)
+         {
+            await channel.DeleteMessageAsync(message);
+         }
+
+         Instance().ClearNotificationMessage(guild.Id, channel.Id);
       }
 
       /// <summary>
@@ -376,7 +416,7 @@ namespace PokeStar.ConnectionInterface
          List<KeyValuePair<string, double>> myList = fuzzy.ToList();
          myList.Sort((pair1, pair2) => pair2.Value.CompareTo(pair1.Value));
          fuzzy = myList.ToDictionary(x => x.Key, x => x.Value);
-         return fuzzy.Keys.Take(NumSuggestions).ToList();
+         return fuzzy.Keys.Take(Global.SELECTION_EMOJIS.Length).ToList();
       }
 
       /// <summary>
@@ -944,12 +984,39 @@ namespace PokeStar.ConnectionInterface
       }
 
       /// <summary>
+      /// Get notification messages for a channel.
+      /// </summary>
+      /// <param name="guild">Id of the guild.</param>
+      /// <param name="channel">Id of the channel.</param>
+      /// <returns>List of message Ids.</returns>
+      public List<ulong> GetNotificationMessages(ulong guild, ulong channel)
+      {
+         string messageStr = NONADBConnector.GetNotificationMessages(guild, channel);
+
+         if (messageStr == null)
+         {
+            return null;
+         }
+
+         string[] messages = messageStr.Split(',');
+         List<ulong> msgList = new List<ulong>();
+         foreach (string msg in messages)
+         {
+            if (!string.IsNullOrEmpty(msg) && !string.IsNullOrWhiteSpace(msg))
+            {
+               msgList.Add(ulong.Parse(msg));
+            }
+         }
+         return msgList;
+      }
+
+      /// <summary>
       /// Updates the notification message for a channel.
       /// </summary>
       /// <param name="guild">Id of the guild.</param>
       /// <param name="channel">Id of the channel.</param>
       /// <param name="message">Id of the message</param>
-      public void UpdateNotificationMessage(ulong guild, ulong channel, ulong message)
+      public void UpdateNotificationMessage(ulong guild, ulong channel, string message)
       {
          NONADBConnector.UpdateNotificationMessage(guild, channel, message);
       }
@@ -961,7 +1028,7 @@ namespace PokeStar.ConnectionInterface
       /// <param name="channel">Id of the channel.</param>
       public void ClearNotificationMessage(ulong guild, ulong channel)
       {
-         NONADBConnector.UpdateNotificationMessage(guild, channel, null);
+         NONADBConnector.UpdateNotificationMessage(guild, channel);
       }
 
       /// <summary>
